+++ src/router.ts
@@ -1,7 +1,9 @@
 /**
  * Provides all routing functionality of a Crea application, based off koa-router.
  */
+import * as koa from 'koa';
 import * as KoaRouter from 'koa-router';
+import { Files, Fields } from 'formidable';
 
 import { Responder } from './responder';
 
@@ -11,8 +13,17 @@ import { Responder } from './responder';
  */
 export class Router extends KoaRouter {}
 
+/** A web application request. */
+export interface Request extends koa.Request {
+  /** Any multipart files for the request. */
+  files?: Files;
+
+  /** Any multipart fields for the request. */
+  fields?: Fields;
+}
+
 /**
- * A Crea-specific implementation of koa-router's router context.
+ * A web application's router context.
  * This will be the router context passed to any middleware in Crea,
  * but it is functionally equivalent to both Koa's default contexts
  * and koa-router's contexts.
@@ -28,10 +39,13 @@ export interface RouterContext extends KoaRouter.IRouterContext {
 
   /** The responder to use for sending REST responses. */
   responder: Responder;
+
+  /** The request for this context. */
+  request: Request;
 }
 
 /**
- * A Crea-specific middleware interface. This will just use
+ * A web application's middleware interface. This will just use
  * our specific router context rather than the default.
  */
 export interface Middleware {